Desktop/CODE/採用確認自動チェック/requirements.txt
selenium>=4.16.0
gspread
requests
chromedriver-binary
oauth2client
pandas
python-dateutil
webdriver-manager>=4.0.1
pytest
numpy